+++ src/index.tsx
@@ -1,8 +1,8 @@
-export { default as Header } from './Header'
-export { default as Content } from './Content'
-export { default as Root, LayoutContext } from './Root'
-export { default as Nav } from './Nav'
-export { default as NavListItem } from './NavListItem'
-export { default as Footer } from './Footer'
-export { default as presets } from './util'
+export * from './Header'
+export * from './Content'
+export * from './Root'
+export * from './Nav'
+export * from './NavListItem'
+export * from './Footer'
+export * from './presets'
 export * from './types'
