+++ stories/Example.tsx
@@ -48,10 +48,7 @@ export const Example: FC<{
   return (
     <ThemeProvider>
       <Root style={{ minHeight: '100vh' }} config={config}>
-        <Header
-          closeMenuIcon={closeMenuIcon}
-          openMenuIcon={openMenuIcon}
-        >
+        <Header closeMenuIcon={closeMenuIcon} openMenuIcon={openMenuIcon}>
           <Typography variant="h5">Application Name</Typography>
           <Box flexGrow={1} />
           <IconButton color="inherit">
